+++ src/pentagram.cpp
@@ -54,10 +54,7 @@ Pentagram::draw_ring(png::Image &_image) const
             if (
                    std::pow(__dx, 2) + std::pow(__dy, 2) <= std::pow(R, 2)
                 && std::pow(__dx, 2) + std::pow(__dy, 2) >= std::pow(r, 2)
-            ) _image.pixel({
-                static_cast<std::uint64_t>(x),
-                static_cast<std::uint64_t>(y)
-            }, this->_color);
+            ) _image.pixel({x, y}, this->_color);
         }
 
 }
